rename SessionStartCloudStorageSecretsModal to CloudStorageSecretsModal
